+++ src/Tests/ContainerApps.fs
@@ -6,15 +6,23 @@ open Farmer.Builders
 open Newtonsoft.Json.Linq
 open Farmer.ContainerApp
 open Farmer.Identity
+open Farmer.Arm
 
-let msi = createUserAssignedIdentity "appUser"
+let msi =
+    createUserAssignedIdentity "appUser"
+
 let containerRegistryName = "myregistry"
 let storageAccountName = "storagename"
 
 let fullContainerAppDeployment =
-    let containerLogs = logAnalytics { name "containerlogs" }
-    let containerRegistryDomain = $"{containerRegistryName}.azurecr.io"
-    let acr = containerRegistry { name containerRegistryName }
+    let containerLogs =
+        logAnalytics { name "containerlogs" }
+
+    let containerRegistryDomain =
+        $"{containerRegistryName}.azurecr.io"
+
+    let acr =
+        containerRegistry { name containerRegistryName }
 
     let storage =
         storageAccount {
@@ -35,7 +43,11 @@ let fullContainerAppDeployment =
                         name "http"
                         add_identity msi
                         active_revision_mode Single
-                        add_registry_credentials [ registry containerRegistryDomain containerRegistryName ]
+
+                        add_registry_credentials
+                            [
+                                registry containerRegistryDomain containerRegistryName
+                            ]
 
                         add_containers
                             [
@@ -65,8 +77,16 @@ let fullContainerAppDeployment =
                         add_http_scale_rule "http-scaler" { ConcurrentRequests = 10 }
                         add_cpu_scale_rule "cpu-scaler" { Utilisation = 50 }
                         add_secret_parameters [ "servicebusconnectionkey" ]
-                        add_env_variables [ "ServiceBusQueueName", "wishrequests" ]
-                        add_secret_expressions [ "containerlogs", containerLogs.PrimarySharedKey ]
+
+                        add_env_variables
+                            [
+                                "ServiceBusQueueName", "wishrequests"
+                            ]
+
+                        add_secret_expressions
+                            [
+                                "containerlogs", containerLogs.PrimarySharedKey
+                            ]
                     }
                     containerApp {
                         name "servicebus"
@@ -88,7 +108,12 @@ let fullContainerAppDeployment =
                                 container {
                                     name "servicebus"
                                     private_docker_image containerRegistryDomain "servicebus" version
-                                    add_volume_mounts [ "empty-v", "/tmp"; "certs-v", "/certs" ]
+
+                                    add_volume_mounts
+                                        [
+                                            "empty-v", "/tmp"
+                                            "certs-v", "/certs"
+                                        ]
                                 }
                             ]
 
@@ -113,12 +138,18 @@ let tests =
     testList
         "Container Apps"
         [
-            let jsonTemplate = fullContainerAppDeployment.Template |> Writer.toJson
+            let jsonTemplate =
+                fullContainerAppDeployment.Template
+                |> Writer.toJson
+
             let jobj = JObject.Parse jsonTemplate
 
             test "Container automatically creates a log analytics workspace" {
-                let env: IBuilder = containerEnvironment { name "testca" }
-                let resources = env.BuildResources Location.NorthEurope
+                let env: IBuilder =
+                    containerEnvironment { name "testca" }
+
+                let resources =
+                    env.BuildResources Location.NorthEurope
 
                 Expect.exists
                     resources
@@ -144,13 +175,19 @@ let tests =
                     |> List.find (fun r -> r.ResourceId.Name.Value = "multienv")
                     :?> Farmer.Arm.App.ContainerApp
 
-                containerApp.EnvironmentVariables.["ServiceBusQueueName"] |> ignore
-                containerApp.EnvironmentVariables.["servicebusconnectionkey"] |> ignore
-                containerApp.EnvironmentVariables.["containerlogs"] |> ignore
+                containerApp.EnvironmentVariables.["ServiceBusQueueName"]
+                |> ignore
+
+                containerApp.EnvironmentVariables.["servicebusconnectionkey"]
+                |> ignore
+
+                containerApp.EnvironmentVariables.["containerlogs"]
+                |> ignore
             }
 
             test "Full container managed environments" {
-                let kubeEnv = jobj.SelectToken("resources[?(@.name=='kubecontainerenv')]")
+                let kubeEnv =
+                    jobj.SelectToken("resources[?(@.name=='kubecontainerenv')]")
 
                 Expect.equal
                     (kubeEnv.["type"] |> string)
@@ -176,13 +213,15 @@ let tests =
                     )
 
                 Expect.equal
-                    (kubeEnvLogAnalyticsCustomerId.["customerId"] |> string)
+                    (kubeEnvLogAnalyticsCustomerId.["customerId"]
+                     |> string)
                     "[reference(resourceId('Microsoft.OperationalInsights/workspaces', 'containerlogs'), '2020-03-01-preview').customerId]"
                     "Incorrect log analytics customerId reference"
             }
 
             test "Full container environment containerApp" {
-                let httpContainerApp = jobj.SelectToken("resources[?(@.name=='http')]")
+                let httpContainerApp =
+                    jobj.SelectToken("resources[?(@.name=='http')]")
 
                 Expect.equal
                     (httpContainerApp.["type"] |> string)
@@ -190,16 +229,28 @@ let tests =
                     "Incorrect type for containerApps"
 
                 Expect.equal (httpContainerApp.["kind"] |> string) "containerapp" "Incorrect kind for containerApps"
-                let ingress = httpContainerApp.SelectToken("properties.configuration.ingress")
-                Expect.isTrue (ingress.SelectToken("external") |> string |> bool.Parse) "Incorrect external ingress"
+
+                let ingress =
+                    httpContainerApp.SelectToken("properties.configuration.ingress")
+
+                Expect.isTrue
+                    (ingress.SelectToken("external")
+                     |> string
+                     |> bool.Parse)
+                    "Incorrect external ingress"
+
                 Expect.equal (ingress.SelectToken("targetPort") |> string |> int) 80 "Incorrect targetPort"
                 Expect.equal (ingress.SelectToken("transport") |> string) "auto" "Incorrect transport"
-                let registries = httpContainerApp.SelectToken("properties.configuration.registries")
+
+                let registries =
+                    httpContainerApp.SelectToken("properties.configuration.registries")
+
                 Expect.hasLength registries 1 "Expected 1 registry"
                 let firstRegistry = registries |> Seq.head
 
                 Expect.equal
-                    (firstRegistry.SelectToken("passwordSecretRef") |> string)
+                    (firstRegistry.SelectToken("passwordSecretRef")
+                     |> string)
                     "myregistry"
                     "Incorrect registry password secretRef"
 
@@ -213,7 +264,9 @@ let tests =
                     "myregistry"
                     "Incorrect registry username"
 
-                let secrets = httpContainerApp.SelectToken("properties.configuration.secrets")
+                let secrets =
+                    httpContainerApp.SelectToken("properties.configuration.secrets")
+
                 Expect.hasLength secrets 2 "Expecting 2 secrets"
                 Expect.equal (secrets.[0].["name"] |> string) "myregistry" "Incorrect name for registry password secret"
 
@@ -223,11 +276,14 @@ let tests =
                     "Incorrect password parameter for registry password secret"
 
                 Expect.equal
-                    (httpContainerApp.SelectToken("properties.managedEnvironmentId") |> string)
+                    (httpContainerApp.SelectToken("properties.managedEnvironmentId")
+                     |> string)
                     "[resourceId('Microsoft.App/managedEnvironments', 'kubecontainerenv')]"
                     "Incorrect kube environment Id"
 
-                let containers = httpContainerApp.SelectToken("properties.template.containers")
+                let containers =
+                    httpContainerApp.SelectToken("properties.template.containers")
+
                 Expect.hasLength containers 1 "Expected 1 http container"
                 let httpContainer = containers |> Seq.head
 
@@ -239,36 +295,48 @@ let tests =
                 Expect.equal (httpContainer.["name"] |> string) "http" "Incorrect container name"
 
                 Expect.equal
-                    (httpContainer.SelectToken("resources.cpu") |> float)
+                    (httpContainer.SelectToken("resources.cpu")
+                     |> float)
                     0.25
                     "Incorrect container cpu resources"
 
                 Expect.equal
-                    (httpContainer.SelectToken("resources.memory") |> string)
+                    (httpContainer.SelectToken("resources.memory")
+                     |> string)
                     "0.50Gi"
                     "Incorrect container memory resources"
 
                 Expect.equal
-                    (httpContainer.SelectToken("resources.ephemeralStorage") |> string)
+                    (httpContainer.SelectToken("resources.ephemeralStorage")
+                     |> string)
                     "1.00Gi"
                     "Incorrect container ephemeral storage resources"
 
-                let scale = httpContainerApp.SelectToken("properties.template.scale")
+                let scale =
+                    httpContainerApp.SelectToken("properties.template.scale")
+
                 Expect.isNotNull scale "properties.scale was null"
                 Expect.equal (scale.["minReplicas"] |> int) 1 "Incorrect min replicas"
                 Expect.equal (scale.["maxReplicas"] |> int) 5 "Incorrect max replicas"
 
-                let serviceBusContainerApp = jobj.SelectToken("resources[?(@.name=='servicebus')]")
-                let volumes = serviceBusContainerApp.SelectToken("properties.template.volumes")
+                let serviceBusContainerApp =
+                    jobj.SelectToken("resources[?(@.name=='servicebus')]")
+
+                let volumes =
+                    serviceBusContainerApp.SelectToken("properties.template.volumes")
+
                 Expect.hasLength volumes 2 "Expecting 2 volumes"
 
                 let serviceBusContainer =
-                    serviceBusContainerApp.SelectToken("properties.template.containers") |> Seq.head
+                    serviceBusContainerApp.SelectToken("properties.template.containers")
+                    |> Seq.head
 
-                let serviceBusVolumeMounts = serviceBusContainer.SelectToken("volumeMounts")
+                let serviceBusVolumeMounts =
+                    serviceBusContainer.SelectToken("volumeMounts")
 
                 Expect.equal
-                    (serviceBusVolumeMounts.[1].["volumeName"] |> string)
+                    (serviceBusVolumeMounts.[1].["volumeName"]
+                     |> string)
                     "empty-v"
                     "Incorrect container volume mount"
 
@@ -278,7 +346,8 @@ let tests =
                     "Incorrect container volume mount"
 
                 Expect.equal
-                    (serviceBusVolumeMounts.[0].["volumeName"] |> string)
+                    (serviceBusVolumeMounts.[0].["volumeName"]
+                     |> string)
                     "certs-v"
                     "Incorrect container volume mount"
 
@@ -336,4 +405,14 @@ let tests =
                              .OkValue)
                     "Container app did not have linked ACR's secret"
             }
+
+            ftest "Turns on Dapr" {
+                let containerApp =
+                    fullContainerAppDeployment.Template.Resources
+                    |> List.find (fun r -> r.ResourceId.Name.Value = "http")
+                    :?> ContainerApp
+
+                Expect.isSome containerApp.DaprConfig "Dapr config was not set"
+            }
+
         ]