+++ src/Farmer/Arm/Network.fs
@@ -45,7 +45,7 @@ let localNetworkGateways =
 let natGateways = ResourceType("Microsoft.Network/natGateways", "2021-08-01")
 
 let privateEndpoints =
-    ResourceType("Microsoft.Network/privateEndpoints", "2020-07-01")
+    ResourceType("Microsoft.Network/privateEndpoints", "2021-05-01")
 
 let virtualNetworkPeering =
     ResourceType("Microsoft.Network/virtualNetworks/virtualNetworkPeerings", "2020-05-01")
@@ -558,6 +558,7 @@ type NetworkInterface =
 
 
 
+
                             |})
                 |}
 
@@ -711,6 +712,7 @@ type PrivateEndpoint =
         Location: Location
         Subnet: SubnetReference
         Resource: LinkedResource
+        CustomNetworkInterfaceName: string option
         GroupIds: string list
     }
 
@@ -725,6 +727,7 @@ type PrivateEndpoint =
                 Location = location
                 Subnet = subnet
                 Resource = Managed resourceId
+                CustomNetworkInterfaceName = None
                 GroupIds = groupIds
             }
             :> IArmResource)
@@ -745,6 +748,7 @@ type PrivateEndpoint =
                 properties =
                     {|
                         subnet = {| id = this.Subnet.ResourceId.Eval() |}
+                        customNetworkInterfaceName = this.CustomNetworkInterfaceName |> Option.toObj
                         privateLinkServiceConnections =
                             [
                                 {|