+++ RELEASE_NOTES.md
@@ -1,8 +1,9 @@
 Release Notes
 =============
 
-## vNext
+## 1.7.7
 * NAT Gateways: Initial support for NAT Gateways.
+* Private Endpoints: Adds `privateEndpoint` builder and option to set custom network interface name.
 
 ## 1.7.6
 * Virtual Machines: Support for adding a VM network interface to a load balancer backend.